+++ src/main/java/edu/tust/neusoft/backend/repository/VisitorRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface VisitorRepository extends JpaRepository<Visitor, Integer> {
-    List<Visitor> findByUserId(int userId);
+    List<Visitor> findByUserId(Long userId);
 }
